events_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
itsdangerous==2.2.0
slowapi==0.1.9
qrcode==7.4.2
ics==0.7.2
pydantic-settings==2.3.4
jinja2==3.1.4
python-dotenv==1.0.1
